remove old edit plan popper
